Refactored the server startup healtchecks, combining it with the PreStartupStoreUpgrader code. Checks and tasks performed before proper server start are now uniformly called PreFlightTasks. These can take arbitrary time to perform (eg. upgrades, recovery).
After the preflight checks, the main server startup uses a timer thread to keep track of startup time. If startup takes longer than the configured timeout,
the startup process is interrupted.

This is all done within the main server process, which means we can remove the code in startup scripts that start a bootstrap JVM separately, as well
as the code that keeps track of startup timeouts.