Fixes ImpermanentDatabase using filesystem (#874)
The main culprit was the StringLogger that created a real log file. Now
a different logging service is used that uses the FileSystemAbstraction
to back the logger file.
Another problem was the KernelDiagnostics - it just created a java.io.File
to check the free and total space, which are 0 for EphemeralFSAbstraction,
and resulted in a division by 0 exception which could be visible in the log.