Attempt at fixing test failures in osx
ImpermanentGraphDatabase is used by RRD tests to create temporary files in it's directory, which lies outside of the control of both maven and the build system. This leads to failures since IGD does not clean up after itself (or before itself), leading to failures because of files already existing. This adds the option to have the store dir of IGD to a known location so these problems are dealt with.